requirements-dev.txt
coverage
isort
astroid
pylint
pylint-django
prospector
django-debug-toolbar
